+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add `--regex-patterns-label` command line option
 
 # 0.3.3
 * Refactor `file_export` and `rich_console`